chatbot/Backend/requirements.txt
flask>=3.0.0
werkzeug>=3.0.0
flask-cors>=4.0.0
pandas>=2.0.0
numpy>=1.24.0
openpyxl>=3.1.0
python-docx>=1.1.0
PyPDF2>=3.0.0
sentence-transformers>=2.7.0
faiss-cpu>=1.7.4
rank-bm25>=0.2.2
transformers>=4.41.0
torch>=2.2.0
scikit-learn>=1.4.0
